fix(v2.1): harden rvr runtime boundaries (#3039)
- trap when rvr reaches an invalid control-flow target or a host
callback rejects its input
- keep rv64 register operands full-width across host I/O
- enable stricter generated C warnings and add boundary regression tests

Resolves INT-8837

+++ extensions/keccak256/rvr/c/rvr_ext_keccak.h
@@ -17,7 +17,7 @@ extern __attribute__((preserve_most)) void rvr_ext_keccakf(
 extern __attribute__((preserve_most)) void rvr_ext_xorin(RvState* state,
                                                          uint64_t buffer_ptr,
                                                          uint64_t input_ptr,
-                                                         uint32_t len,
+                                                         uint64_t len,
                                                          uint32_t chip_idx);
 
 #endif /* RVR_EXT_KECCAK_H */
